requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
fastapi==0.121.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
python-dotenv==1.2.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
